temp

+++ app/src/main/java/br/com/if91wallet/vo/CardVo.kt
@@ -1,7 +1,7 @@
 package br.com.if91wallet.vo
 
 data class CardVo(
-    val cardNumber : Int,
+    val cardNumber : Long,
     val name : String,
     val expirationDate : Int,
     val cvv : String
